Fix for handling archived projects on Redmine
Note: When an archived project on Redmine is un-archived, the synchronization with CAST would still be successful. However, the issues assoicated with the un-archived projects would not show up on CAST as synchronization function only fetches the issues updated after recent entry on 'lastUpdatedTime' column of Issues table. The capabilities associated with an un-archived project  would also be deleted.